pyproject.toml
[project]
name = "typeandlearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "nltk>=3.9.2",
    "google-genai>=1.62.0",
    "pandas>=3.0.0",
    "tabulate>=0.9.0",
]
